% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{allWithin}
\alias{allWithin}
\title{Check if all elements of a numeric vector fall within a tolerable range.}
\usage{
allWithin(x, tol = 0, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{tol}{Tolerable range. Default = 0 to check if all elements are equal.}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds. Default = FALSE}
}
\description{
Check if all elements of a numeric vector fall within a tolerable range.
}

