% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{appendDemo}
\alias{appendDemo}
\title{Append the 'demo' data frames from one phd data set onto another phd data
set.}
\usage{
appendDemo(phd.without, phd.with)
}
\arguments{
\item{phd.without}{Phd data set without the demo data frames.}

\item{phd.with}{Phd data set with the demo data frames}
}
\description{
Append the 'demo' data frames from one phd data set onto another phd data
set.
}

