% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variables.R
\name{compute}
\alias{compute}
\title{Compute a set of variables for each participant.}
\usage{
compute(x, vars = c("accuracy", "confidence", "bias", "discrimination",
  "post"), bind = TRUE, na.rm = TRUE)
}
\arguments{
\item{x}{Data frame from phd sample.}

\item{vars}{Character vector of variables to be computed. Default/Allowed =
c("accuracy", "confidence", "bias", "discrimination", "post")}

\item{bind}{logical. Should variable results be bound to original data frame x? Default = TRUE.}

\item{na.rm}{logical. Should missing values (including NaN) be omitted from
the calculations? Default = True}
}
\description{
Computes a set of variables for each participant in a Jackson PhD data frame.
}

