% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variables.R
\name{computeBias}
\alias{computeBias}
\title{Compute bias (over/underconfidence) for each participant.}
\usage{
computeBias(x, na.rm = TRUE)
}
\arguments{
\item{x}{Data frame from phd sample.}

\item{na.rm}{logical. Should missing values (including NaN) be omitted from the calculations? Default = True}
}
\value{
Bias vector.
}
\description{
Compute the bias score for each participant in the data frame.
To be called with compute().
}
\examples{
computeBias(phd[[1]]$EA)
}

