% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variables.R
\name{computeConfidence}
\alias{computeConfidence}
\title{Compute mean confidence as a percentage for each participant in Jackson Phd test data frame.}
\usage{
computeConfidence(x, na.rm = TRUE)
}
\arguments{
\item{x}{Data frame from phd sample.}

\item{na.rm}{logical. Should missing values (including NaN) be omitted from the calculations? Default = True}
}
\value{
Confidence vector.
}
\description{
Compute the mean of all 'c' columns (scored 1 for correct and 0 for
incorrect) for each participant (row).
}
\examples{
computeConfidence(phd[[1]]$EA)
}

