% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variables.R
\name{computeDecisionCat}
\alias{computeDecisionCat}
\title{Compute percentage of decisions of a particular category.}
\usage{
computeDecisionCat(x, a, d, group = c(T))
}
\arguments{
\item{a}{0 or 1 indicating whether accuracy should be 0 or 1.}

\item{group}{Logical vector used to return a group of the candidate results.
See ?getCol for details.}

\item{Data}{frame from phd sample.}
}
\description{
Compute the percentage of decisions in a given decision category. E.g., a = 1
and d = 0 returns percentage of misses. To be called with compute()
}

