% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variables.R
\name{computeDiscrimination}
\alias{computeDiscrimination}
\title{Compute discrimination (confidence for correct answers - confidence for
incorrect answers) for each participant in Jackson Phd test data frame.
To be called with compute().}
\usage{
computeDiscrimination(x, group = c(T), na.rm = TRUE)
}
\arguments{
\item{x}{Data frame from phd sample.}

\item{group}{Logical vector used to return a group of the candidate results.
See ?getCol for details.}

\item{na.rm}{logical. Should missing values (including NaN) be omitted from the calculations? Default = True}
}
\value{
Discrimination vector.
}
\description{
Compute discrimination (confidence for correct answers - confidence for
incorrect answers) for each participant in Jackson Phd test data frame.
To be called with compute().
}
\examples{
computeDiscrimination(phd[[1]]$EA)
}

