% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variables.R
\name{computePost}
\alias{computePost}
\title{Compute the POST (Point Of Sufficient cerTainty) for each participant in Jackson Phd test data frame.}
\usage{
computePost(x)
}
\arguments{
\item{x}{Data frame from phd sample.}

\item{na.rm}{logical. Should missing values (including NaN) be omitted from
the calculations? Default = True}
}
\value{
POST vector. Any values < 0 or > 100 scored as NA.
}
\description{
Compute the POST (Point Of Sufficient cerTainty) for each participant in Jackson Phd test data frame.
}
\examples{
computePost(phd[[1]]$EA)
}

