% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getCol}
\alias{getCol}
\title{Get item-level column names from Jackson Phd test data frame.}
\usage{
getCol(x, i, group = c(T))
}
\arguments{
\item{x}{Data frame from phd sample.}

\item{i}{Character indicating information type to get. 'a' = accuracy, 'c' =
confidence, 'd' = decision, 't' = time.}

\item{group}{Logical vector used to return a group of the candidate results.
Default = c(TRUE) to return all results.}
}
\value{
Vector of column names.
}
\description{
Get item-level column names for answer accuracy (a), confidence (c), decision (d)
or time (t) from data frame with regex call.
}
\examples{
getCol(phd[[1]]$EA, 'a')
getCol(phd[[1]]$EA, "a|c")
getCol(phd[[1]]$EA, "[ac]")
getCol(phd[[1]]$EA, 'a', c(T, F))  # return off columns
}

