% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{nItem}
\alias{nItem}
\title{Return the number of test items for a given Jackson Phd test data frame.}
\usage{
nItem(x)
}
\arguments{
\item{x}{Data frame from phd sample.}
}
\value{
Vector of column names.
}
\description{
Queries and returns the number of accuracy columns.
}
\examples{
nItem(phd[[1]]$EA)
}

