% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{phd}
\alias{phd}
\title{Data from Simon A Jackson's PhD.}
\format{A list with four lists for each of four samples. Within each sample
  list, there are multiple data frames - one for for each test that was
  admininistered.}
\usage{
phd
}
\description{
A list of datasets containing results on various test for four samples
collected for research by Simon A Jackson. Data list available as 'phd'.
Additional helper functions provided.
}
\keyword{datasets}

