% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{phdApply}
\alias{phdApply}
\title{Apply a function to all data frames in a phd data set
(or copy of it).}
\usage{
phdApply(phd.dat, FUN, ..., demo.rm = TRUE)
}
\arguments{
\item{phd.dat}{Data set on which to apply function. Either phd or copy of it.}

\item{FUN}{the function to be applied}

\item{...}{optional arguments to FUN.}

\item{demo.rm}{a logical value indicating whether the demographic data frames
should be removed before the analysis proceeds.}
}
\description{
This is a convenience function to apply a function (FUN) to all data frames
within the phd data set (or copy of it).
}
\examples{
x <- phdApply(phd, compute)
sapply(x, names)
head(x[[1]]$EA)

x <- phdApply(phd, compute, "accuracy")
x <- appendDemo(x, phd)  # append demographics back on
lapply(x, names)
head(x[[1]]$EA)
}

