% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{removeDemo}
\alias{removeDemo}
\title{Remove the 'demo' data frames from a phd data set.}
\usage{
removeDemo(phd.dat)
}
\arguments{
\item{phd.dat}{Data set of phd data.}
}
\description{
Remove the 'demo' data frames from a phd data set.
}

