% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variables.R
\name{computeAccuracy}
\alias{computeAccuracy}
\title{Compute mean accuracy as a percentage for each participant in Jackson Phd test data frame.}
\usage{
computeAccuracy(x, na.rm = TRUE)
}
\arguments{
\item{x}{Data frame from phd sample.}

\item{na.rm}{logical. Should missing values (including NaN) be omitted from the calculations? Default = True}
}
\value{
Accuracy vector.
}
\description{
Compute the mean of all 'a' columns (scored 1 for correct and 0 for
incorrect) for each participant (row) and multiply by 100.
To be called with compute().
}
\examples{
computeAccuracy(phd[[1]]$EA)
}

