% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getCol}
\alias{getCol}
\title{Get item-level column names from Jackson Phd test data frame.}
\usage{
getCol(x, i)
}
\arguments{
\item{x}{Data frame to get columns indexes from.}

\item{i}{Character indicating information type to get. 'a' = accuracy, 'c' =
confidence, 'd' = decision, 't' = time.}
}
\value{
Vector of column names.
}
\description{
Get item-level column names for answer accuracy (a), confidence (c), decision (d)
or time (t) from data frame.
}

