% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{phdApply}
\alias{phdApply}
\title{Apply a function to all data frames (except demographics) in the phd data set
(or copy of it).}
\usage{
phdApply(all, FUN, ...)
}
\arguments{
\item{all}{Data set on which to apply function. Either phd or copy of it.}

\item{FUN}{the function to be applied}

\item{...}{optional arguments to FUN.}
}
\description{
This is a convenience function to apply a function (FUN) to all data frames
within the phd data set (or copy of it).
}

