% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{stretchPhd}
\alias{stretchPhd}
\title{Convert data frame from phd sample to long format.}
\usage{
stretchPhd(x, demo = NULL)
}
\arguments{
\item{x}{Data frame from phd sample.}

\item{demo}{Matching demographics data frame from phd sample. If NULL, no
demographics will be added.}
}
\description{
Convert data frame from phd sample to long format.
}

